# Back_end reservation/openapi.yaml
openapi: 3.0.3
info:
  title: API de Reservas
  version: 1.0.0
  description: API para la gestion de usuarios, eventos y reservaciones.
  license:
    name: ISC
    url: https://opensource.org/license/isc-license-txt
servers:
  - url: http://localhost:3000
    description: Servidor local

tags:
  - name: Auth
    description: Registro, autenticacion y perfil de usuario
  - name: Eventos
    description: Consulta y administracion de eventos
  - name: Reservaciones
    description: Creacion y consulta de reservaciones

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserCredentials:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: usuario1
        password:
          type: string
          format: password
          example: clave123
        type_user:
          type: integer
          description: 1 para administrador y 2 para usuario normal.
          enum: [1, 2]
          example: 2
    LoginCredentials:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin_2026
        password:
          type: string
          format: password
          example: admin2026
    EventInput:
      type: object
      required:
        - name
        - description
        - location
        - date
        - hour
        - total_tickets
        - available_tickets
      properties:
        name:
          type: string
          example: Concierto Rock
        description:
          type: string
          example: Concierto en vivo
        location:
          type: string
          example: Ciudad de Mexico
        date:
          type: string
          format: date
          example: '2026-12-01'
        hour:
          type: string
          example: '20:00:00'
        total_tickets:
          type: integer
          minimum: 0
          example: 100
        available_tickets:
          type: integer
          minimum: 0
          example: 50
    ReservationInput:
      type: object
      required:
        - id_event
        - date_reserv
        - amount
      properties:
        id_event:
          type: integer
          example: 7
        date_reserv:
          type: string
          format: date
          example: '2026-11-05'
        amount:
          type: integer
          minimum: 1
          example: 3
    Error:
      type: object
      properties:
        message:
          type: string
          example: Error en la solicitud

paths:
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Registrar un nuevo usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
      responses:
        '201':
          description: Usuario creado correctamente.
        '400':
          description: Datos de registro invalidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Iniciar sesion y obtener un token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
      responses:
        '200':
          description: Inicio de sesion exitoso.
        '401':
          description: Credenciales invalidas.

  /api/auth/profile:
    get:
      tags: [Auth]
      operationId: getProfile
      summary: Obtener el perfil del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del usuario.
        '401':
          description: Token no proporcionado o invalido.

  /api/events:
    get:
      tags: [Eventos]
      operationId: listEvents
      summary: Obtener todos los eventos
      security: []
      responses:
        '200':
          description: Lista de eventos.
        '400':
          description: Solicitud invalida.
    post:
      tags: [Eventos]
      operationId: createEvent
      summary: Crear un evento
      description: Requiere autenticacion y rol de administrador.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '201':
          description: Evento creado correctamente.
        '401':
          description: Token no proporcionado o invalido.
        '403':
          description: El usuario no tiene permisos de administrador.

  /api/events/filter:
    get:
      tags: [Eventos]
      operationId: filterEvents
      summary: Filtrar eventos
      security: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
          example: '2026-12-01'
        - name: hour
          in: query
          schema:
            type: string
          example: '20:00:00'
        - name: location
          in: query
          schema:
            type: string
          example: Ciudad de Mexico
      responses:
        '200':
          description: Lista de eventos filtrados.
        '400':
          description: Parametros de filtrado invalidos.
        '500':
          description: Error al filtrar eventos.

  /api/events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        example: 7
    get:
      tags: [Eventos]
      operationId: getEventById
      summary: Obtener un evento por su ID
      security: []
      responses:
        '200':
          description: Evento encontrado.
        '400':
          description: ID de evento invalido.
        '404':
          description: Evento no encontrado.
    put:
      tags: [Eventos]
      operationId: updateEvent
      summary: Actualizar un evento
      description: Requiere autenticacion y rol de administrador.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '200':
          description: Evento actualizado correctamente.
        '401':
          description: Token no proporcionado o invalido.
        '403':
          description: El usuario no tiene permisos de administrador.
        '404':
          description: Evento no encontrado.
    delete:
      tags: [Eventos]
      operationId: deleteEvent
      summary: Eliminar un evento
      description: Requiere autenticacion y rol de administrador.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Evento eliminado correctamente.
        '401':
          description: Token no proporcionado o invalido.
        '403':
          description: El usuario no tiene permisos de administrador.
        '404':
          description: Evento no encontrado.

  /api/reservas:
    get:
      tags: [Reservaciones]
      operationId: checkReservationsRoute
      summary: Comprobar la ruta de reservas
      security: []
      responses:
        '200':
          description: Mensaje de disponibilidad de la ruta.
        '400':
          description: Solicitud invalida.

  /api/reservations:
    get:
      tags: [Reservaciones]
      operationId: listReservations
      summary: Obtener todas las reservaciones
      security: []
      responses:
        '200':
          description: Lista de reservaciones.
        '400':
          description: Solicitud invalida.
    post:
      tags: [Reservaciones]
      operationId: createReservation
      summary: Crear una reservacion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationInput'
      responses:
        '201':
          description: Reservacion creada y boletos descontados.
        '400':
          description: Datos de reservacion invalidos.
        '401':
          description: Token no proporcionado o invalido.
        '404':
          description: Evento no encontrado.
        '409':
          description: No hay suficientes boletos disponibles.

  /api/reservations/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
        example: 3
    get:
      tags: [Reservaciones]
      operationId: listReservationsByUser
      summary: Obtener las reservaciones de un usuario
      security: []
      responses:
        '200':
          description: Lista de reservaciones del usuario.
        '400':
          description: ID de usuario invalido.
        '404':
          description: Usuario no encontrado.